aggregating sigs WIP
